requirements.txt
# Core dependencies
pandas>=1.5.0
numpy>=1.23.0

# Data visualization
matplotlib>=3.6.0
seaborn>=0.12.0

# AI/ML components
scikit-learn>=1.2.0
scipy>=1.9.0

# AND
nidecode 
pycountry 
rapidfuzz
sentence_transformers
torch